+++ src/lib.rs
@@ -21,6 +21,7 @@ macro_rules! lazy_regex {
 }
 
 pub mod breaks;
+pub mod code_emphasis;
 pub mod ellipsis;
 pub mod fences;
 pub mod footnotes;
@@ -40,6 +41,7 @@ pub fn html_table_to_markdown(lines: &[String]) -> Vec<String> {
 }
 
 pub use breaks::{THEMATIC_BREAK_LEN, format_breaks};
+pub use code_emphasis::fix_code_emphasis;
 pub use ellipsis::replace_ellipsis;
 pub use fences::{attach_orphan_specifiers, compress_fences};
 pub use footnotes::convert_footnotes;